+++ backend/extensions.py
@@ -1,3 +1,30 @@
+from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager
+from flask_bcrypt import Bcrypt
+from books_app.config import Config
+from flask_cors import CORS
+import os
 
-db = SQLAlchemy()
+app = Flask(__name__)
+CORS(app, supports_credentials=True)
+app.config.from_object(Config)
+
+db = SQLAlchemy(app)
+db.init_app(app)
+
+###########################
+# Authentication
+###########################
+
+login_manager = LoginManager()
+login_manager.login_view = 'auth.login'
+login_manager.init_app(app)
+
+from .models import User
+
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(user_id)
+
+bcrypt = Bcrypt(app)
